Move precision helper to utils to avoid unnecessary imports
