+++ nemo/export/trt_llm/converter/model_to_trt_llm_ckpt.py
@@ -17,13 +17,13 @@ import logging
 import multiprocessing
 from collections import defaultdict
 from pathlib import Path
-from typing import Optional, Union
 
 import torch
 from tensorrt_llm._utils import pad_vocab_size, str_dtype_to_torch
 from tqdm import tqdm
 
 from nemo.export.trt_llm.converter.utils import save_scaling_factor, save_val, split_and_save_weight, weights_dict
+from nemo.export.utils import torch_dtype_from_precision
 
 LOGGER = logging.getLogger("NeMo")
 
@@ -34,22 +34,6 @@ layer_names = {
     "final_layernorm.weight": "final_layernorm.weight",
     "final_layernorm.bias": "final_layernorm.bias",
 }
-
-
-def torch_dtype_from_precision(precision: Union[int, str], megatron_amp_O2: Optional[bool] = None) -> torch.dtype:
-    """Mapping from PTL precision types to corresponding PyTorch parameter datatype."""
-    # Copied from nemo.collections.nlp.parts.utils_funcs to avoid extra depenencies for NIM.
-    if megatron_amp_O2 is not None and megatron_amp_O2 is False:
-        return torch.float32
-
-    if precision in ['bf16', 'bf16-mixed']:
-        return torch.bfloat16
-    elif precision in [16, '16', '16-mixed']:
-        return torch.float16
-    elif precision in [32, '32', '32-true']:
-        return torch.float32
-    else:
-        raise ValueError(f"Could not parse the precision of `{precision}` to a valid torch.dtype")
 
 
 def extract_layers_with_prefix(model_, prefix):
